backend/requirements.txt
# E-Charge 聚合平台毕业设计 - 精简依赖
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
pydantic==2.12.5
python-dotenv==1.2.1
alembic==1.12.1
pydantic-settings